留言的创建和删除

+++ models/posts.js
@@ -1,5 +1,6 @@
 const marked = require('marked')
 const Post = require('../lib/mongo').Post
+const CommentModel = require('./comments')
 
 // 将 post 的 content 从 markdown 转换成 html
 // contentToHtml注册在PostModel上，只针对PostModel有效
@@ -16,6 +17,29 @@ Post.plugin('contentToHtml', {
   }
 })
 
+// 给 post 添加留言数 commentsCount
+Post.plugin('addCommentsCount', {
+  afterFind (posts) {
+    return Promise.all(posts.map((post) => {
+      return CommentModel.getCommentsCount(post._id)
+        .then((commentsCount) => {
+          post.commentsCount = commentsCount
+          return post
+        })
+    }))
+  },
+  afterFindOne (post) {
+    if (post) {
+      return CommentModel.getCommentsCount(post._id)
+        .then((count) => {
+          post.commentsCount = count
+          return post
+        })
+    }
+    return post
+  }
+})
+
 module.exports = {
   // 创建一篇文章
   create: function (post) {
@@ -27,6 +51,7 @@ module.exports = {
       .findOne({ _id: postId })
       .populate({ path: 'author', model: 'User' })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec()
   },
@@ -41,6 +66,7 @@ module.exports = {
       .populate({ path: 'author', model: 'User' })
       .sort({ _id: -1 })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec()
   },
@@ -61,8 +87,14 @@ module.exports = {
   updatePostById (postId, data) {
     return Post.update({ _id: postId }, { $set: data }).exec()
   },
-  // 通过文章 id 删除文章
-  deletePostById (postId) {
-    return Post.deleteOne({ _id: postId }).exec()
+  // 通过用户 id 和 文章 id 删除文章
+  deletePostById (author, postId) {
+    return Post.deleteOne({ author, _id: postId }).exec()
+      .then((res) => {
+        // 文章删除后，再删除该文章下的所有留言
+        if (res.result.ok && res.result.n > 0) {
+          return CommentModel.delCommentsByPostId(postId)
+        }
+      })
   }
 }
